+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -285,7 +285,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K94HF0G79K7S0JV7CY053VAY"),
+  _id: z.string().default("01K94ZA4Y2843MV4HSXZEHZN1K"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -329,7 +329,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01K94HF0G79K7S0JV7CY053VAY"),
+  id: z.string().default("01K94ZA4Y2843MV4HSXZEHZN1K"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,